Refactor the display options
