+++ app.rb
@@ -2,12 +2,11 @@ require_relative 'book'
 require_relative 'teacher'
 require_relative 'student'
 require_relative 'rental'
+require_relative 'display'
 
 class App
   def initialize
     @books = []
-    # @teacher = []
-    # @student = []
     @people = []
     @rental = []
   end
@@ -124,20 +123,20 @@ class App
     end
   end
 
-  def display_options
-    puts 'Please choose an option by entering a number'
-    puts '1. List all books.'
-    puts '2. List all people.'
-    puts '3. Create a person.'
-    puts '4. Create a book.'
-    puts '5. Create a rental.'
-    puts '6. List all rentals.'
-    puts '7. Exit app'
+  # def display_options
+  #   puts 'Please choose an option by entering a number'
+  #   puts '1. List all books.'
+  #   puts '2. List all people.'
+  #   puts '3. Create a person.'
+  #   puts '4. Create a book.'
+  #   puts '5. Create a rental.'
+  #   puts '6. List all rentals.'
+  #   puts '7. Exit app'
 
-    option_id = gets.chomp
-    puts option_id
-    run_app(option_id)
-  end
+  #   option_id = gets.chomp
+  #   puts option_id
+  #   run_app(option_id)
+  # end
 
   def exit_app
     puts 'Thank you for using this app! '
